Wed, Aug 21, 2024, 9:28 PM -04:00

+++ week2/multi.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 // Function prototype
-float multiply(float a, float b);
+float multiply(int a, int b);
 
 int main(void)
 {
@@ -18,7 +18,7 @@ int main(void)
 }
 
 // Function definition
-float multiply(float a, float b)
+int multiply(int a, int b)
 {
     return a * b;
 }
